% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geTMMnorm.R
\name{geTMMnorm}
\alias{geTMMnorm}
\title{Performs Genewise Trimmed Means of M-values normalisation of RNAseq read counts}
\usage{
geTMMnorm(
  counts = counts,
  species = NULL,
  expname = "norm",
  meta = meta,
  id = "ENSEMBL"
)
}
\arguments{
\item{counts}{data frame count/mRNA expression matrix with samples in columns and genes in rows}

\item{species}{character accepts either mouse or human}

\item{expname}{character name of the experiment}

\item{id}{character gene id type}
}
\value{
data frame of normalised mRNA counts
}
\description{
Essy geTMM normalisation
}
\details{
ceeb
}
\examples{
geTMMnorm()

}
