% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOrthologs.R
\name{getOrthologs}
\alias{getOrthologs}
\title{Converts human genes to mouse orthologs, and vice-versa}
\usage{
getOrthologs(
  geneIDs,
  from = c("mouse", "human"),
  to = c("human", "mouse"),
  identifier = "entrezgene_id"
)
}
\arguments{
\item{geneIDs}{character vector of gene IDs}

\item{from}{organism from which to convert gene IDs}

\item{to}{organism to which to convert geneIDs}

\item{identifier}{type of gene identifier}
}
\value{
Data frame of mapped gene IDs and annotations
}
\description{
Orthologs converter
}
\details{
ceeb
}
\examples{
getOrthologs()

}
