% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onePath.R
\name{lazyPaths}
\alias{lazyPaths}
\title{Pipleine for Geneset Enrichement and Pathway analysis}
\usage{
lazyPaths(
  sigDEGs = NULL,
  mods = NULL,
  species = NULL,
  orthogenes = NULL,
  orthospecies = NULL,
  setRank = NULL,
  CTDbase = NULL,
  chemDis = NULL,
  chemIX = NULL,
  msigDB = NULL,
  msigCat = NULL,
  minGenes = 50,
  useRank = TRUE,
  geneSetPCut = 0.01,
  geneSetFDRCut = 0.05,
  ppiGSEA = NULL,
  clusterGS = NULL,
  cytoscape = NULL,
  expname = NULL,
  seed = NULL,
  directory = NULL,
  cores = NULL
)
}
\arguments{
\item{mods}{list of WGCNA modules outputed from oneNet function}

\item{species}{character, accepts mouse and human}

\item{orthogenes}{Data frame output of 'getOrthologs' function}

\item{orthospecies}{character, takes 'mouse' or 'human'}

\item{setRank}{logical, perform setRank GSEA?}

\item{CTDbase}{logical, include Chemical and Toxigenomics Data base in GSEA?}

\item{chemDis}{logical, if CTDbase, include chemical-gene-disease interactions in GSEA?}

\item{chemIX}{logical, if CTDbase, include chemical-gene interactions in GSEA?}

\item{msigDB}{logical, include msigDB in GSEA?}

\item{msigCat}{list of desired msigDB sub-categories}

\item{ppiGSEA}{logical, map GSEA results to protein-protein interaction network?}

\item{clusterGS}{logical, perform hierarchal clustering on GSEA results?}

\item{cytoscape}{logical, export networks to cytoscape?}

\item{expname}{character, name of experiment}

\item{seed}{integer, random seed}

\item{directory}{character, project path}

\item{cores}{integer, number of cpu cores to use}

\item{sigDEG}{list of data frames output of lazyDGE function or DE = 'noiseq' of lazySeq function}
}
\description{
Leverages the wonferful setRank and gsCluster packages for GSE and Pathway analysis
}
\details{
ceeb
}
\examples{
onePath()

}
