% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matcha.R
\name{matcha}
\alias{matcha}
\title{Match sample order between count matrix and metadata}
\usage{
matcha(counts = counts, meta = meta)
}
\arguments{
\item{counts}{data frame, counts matrix}

\item{meta}{data frame, sample information}
}
\value{
data frames with sample names in right order
}
\description{
Match sample order between count matrix and metadata
}
\details{
Ceeb
}
\examples{
matcha()

}
