% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneNet.R
\name{oneNet}
\alias{oneNet}
\title{Automated WGCNA analysis pipeline for Bulk RNAseq data}
\usage{
oneNet(
  counts = counts,
  meta = meta,
  species = NULL,
  directory = getwd(),
  normalise = NULL,
  expname = "oneNet_exp",
  seed = 42069,
  cores = NULL,
  includeZeros = FALSE,
  CPMfilter = NULL,
  CPMcutoff = NULL,
  dropOut = NULL,
  batch = NULL,
  svaCor = NULL,
  svaCovar = NULL,
  conCovar = NULL,
  svaFormula = NULL,
  RsqCut = NULL,
  normFormula = NULL,
  choosePower = NULL,
  corFunc = "bicor",
  meanExprStep = 0.01,
  netType = "signed",
  spQN = NULL,
  spQNplots = FALSE,
  QCplots = NULL,
  wgcna = NULL,
  cutDepth = 3,
  kMEmerge = NULL,
  kMEclustThresh = 50,
  kMEclustMrgPcnt = 50,
  MergeCutHeight = 0.2
)
}
\arguments{
\item{counts}{data frame  Raw bulk mRNA seq counts genes as rows, columns as samples.}

\item{meta}{data frame Metadata table with samples as rows and variables as columns. Compulsory columns include 'id' with sample names corresponding to those in counts, 'condition' with relevant tests and controls, 'priority' with integers starting from 0 to 1 for indicating preferred direction of pairwise comparison and 'class' indicating whether the sample is a control or disease/test.}

\item{species}{character Takes 'mouse' or 'human'.}

\item{directory}{character Name of working folder.}

\item{normalise}{logical, character If TRUE, raw counts will be geTMM normalised. If "CQN", raw counts will be quantile normalised}

\item{expname}{character Name of expreriment. E.g., "Exp1_DGEA"}

\item{seed}{numeric Seed from which randomness grows.}

\item{cores}{numeric Specify number of CPU cores.}

\item{includeZeros}{logical If TRUE. zero values will be nudged by 1x10^-16 to facilitate PCA. Otherwise, rows with zero values will be removed.}

\item{CPMfilter}{logical If TRUE, raw counts will be filtered based of counts per milion (CPM) cutoff}

\item{CPMcutoff}{numeric Cutoff for CPM filtering e.g., 2}

\item{dropOut}{logical If TRUE, outliers will be dropped based on PCA distance.}

\item{batch}{logical If TRUE, data will undergo batch correction}

\item{svaCor}{logical If TRUE, performs surrogatew variable analysis and correction.}

\item{svaCovar}{character Covariates of interest whose influence should not be removed e.g., condition}

\item{conCovar}{character Potentially confounding covariates whose influence might want to be removed e.g., gender}

\item{svaFormula}{character Formula for SVA e.g. id ~ condition + gender}

\item{RsqCut}{numeric Desired Rsquared cutoff for scale-free fit e.g, 0.9}

\item{choosePower}{logical If TRUE, will attempt to choose integer exponent which results in scale-free gene expression covariance network. If initial correlation network is not scale free, will iterativelty remove genes based on mean expression cutoff. If "bicor" algorithm is chosen initially and the function fails to produce a scale free network with R^2 > 0.75,}

\item{corFunc}{character The chosen correlation algorithm e.g., "bicor" for bi-correlation or "cor" for pearson's correlation. See WGCNA vignette for more details.}

\item{meanExprStep}{numeric If ChoosePower is TRUE and the intial correlation network is not scale free, the percentage of mean expression cutoff at which genes are shaved off at each iteration.}

\item{netType}{character The type of correlation network e.g., "signed", "unsigned".}

\item{spQN}{logical If TRUE, performs spatial quantile normalisation on scale-free correlation network. Attempt to correlation bias of higher expressed genes.}

\item{spQNplots}{logical If TRUE, plots all quality control plots from spQN}

\item{QCplots}{logical If TRUE, plots all quality control plots associated with analysis}

\item{wgcna}{logical If TRUE, performs weighted gene co-expression neteork analysis. See WGCNA vigette for more details.}

\item{cutDepth}{numeric Depth at which to cut heirarchal clustering tree default: 3. See WGCNA vigette for more details.}

\item{kMEmerge}{logical If TRUE, merges clusters based on gene module membership. See WGCNA vigette for more details.}

\item{kMEclustMrgPcnt}{numeric The percentage of genes above the threshold percentage in a given module e.g., 75. See WGCNA vigette for more details.}

\item{orthologs}{logical If TRUE, maps genes to orthologs of given species and continues analysis with these orthologous genes.}

\item{orthogenes}{character list List of character vector of orthologous ENRTEZID genes (mouse or human)}

\item{orthospecies}{character Takes 'mouse' or 'human'}

\item{normFormu}{character Formula used for variance stabilising normalisation e.g., conditon ~ id}

\item{kMEclustThres}{numeric The percentage genes of modules to be included for merging e.g., 50. See WGCNA vigette for more details.}
}
\value{
List of WGCNA modules
}
\description{
Run and forget WGCNA analysis
}
\details{
ceeb
}
\examples{
oneNet()

}
