% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quikVolcano.R
\name{quikVolcano}
\alias{quikVolcano}
\title{Generate volcano plots from differentially expressed genes table}
\usage{
quikVolcano(
  DEGList = NULL,
  id = NULL,
  Log2FC = NULL,
  pval = NULL,
  cutoff = 0.05,
  expname = NULL
)
}
\arguments{
\item{id}{character, name of symbol column}

\item{Log2FC}{character, name of log fold change column}

\item{pval}{character, name of pval column}

\item{cutoff}{p-value cutoff}

\item{expname}{name of experiment}

\item{DEGlist}{list of data frames output of lazyDGE function or DE = 'noiseq' of lazySeq function}
}
\value{
pdf with volcano plots
}
\description{
Simple function for plotting volcanoes in base R
}
\details{
ceeb
}
\examples{
quikVolcano()

}
